Followup on previous double to int cleanups.

+++ source/cpu/cHardwareCPU.cc
@@ -3771,7 +3771,7 @@ bool cHardwareCPU::Inst_SenseOpinionResourceQuantity(cAvidaContext& ctx)
   // check if this is a valid group
   if(m_organism->HasOpinion()) {
     int opinion = m_organism->GetOpinion().first;
-    int res_opinion = res_count[opinion] * 100 + 0.5;
+    int res_opinion = int(res_count[opinion] * 100 + 0.5);
     int reg_to_set = FindModifiedRegister(REG_BX);
     GetRegister(reg_to_set) = res_opinion;
   }
@@ -3786,7 +3786,7 @@ bool cHardwareCPU::Inst_SenseDiffFaced(cAvidaContext& ctx)
     int reg_to_set = FindModifiedRegister(REG_BX);
     double faced_res = m_organism->GetOrgInterface().GetFacedCellResources(ctx)[opinion];  
     // return % change
-    int res_diff = (faced_res - res_count[opinion])/res_count[opinion] * 100 + 0.5;
+    int res_diff = int((faced_res - res_count[opinion])/res_count[opinion] * 100 + 0.5);
     GetRegister(reg_to_set) = res_diff;
   }
   return true;
@@ -5561,7 +5561,7 @@ bool cHardwareCPU::Inst_RotateUphill(cAvidaContext& ctx)
     }
   }
   // return % change
-  int res_diff = (max_res - current_res[opinion])/current_res[opinion] * 100 + 0.5;
+  int res_diff = int((max_res - current_res[opinion])/current_res[opinion] * 100 + 0.5);
   int reg_to_set = FindModifiedRegister(REG_BX);
   GetRegister(reg_to_set) = res_diff;
   return true;
@@ -8483,7 +8483,7 @@ bool cHardwareCPU::Inst_ReadFacedCellData(cAvidaContext& ctx)
   assert(m_organism != 0);
   const int out_reg = FindModifiedRegister(REG_BX);
   // return % diff (FacedCellData is already int)
-  int my_vit = m_organism->GetVitality() + 0.5;
+  int my_vit = int(m_organism->GetVitality() + 0.5);
   int vit_diff = (m_organism->GetFacedCellData() - my_vit)/my_vit * 100;
   GetRegister(out_reg) = vit_diff;
 
@@ -8523,7 +8523,7 @@ bool cHardwareCPU::Inst_GetResStored(cAvidaContext& ctx)
   const int out_reg = FindModifiedRegister(REG_BX);
   const int resource = m_world->GetConfig().COLLECT_SPECIFIC_RESOURCE.Get();
   // needs to return int...we round down so that they don't think there is more available than they need
-  int res_stored = m_organism->GetRBin(resource) * 100 - 0.5;
+  int res_stored = int(m_organism->GetRBin(resource) * 100 - 0.5);
   GetRegister(out_reg) = res_stored;
   return true;
 }
@@ -8532,7 +8532,7 @@ bool cHardwareCPU::Inst_MarkCellWithVitality(cAvidaContext& ctx)
 {
   assert(m_organism != 0);
   // SetCellData() needs to be int
-  int my_vit = m_organism->GetVitality() + 0.5;
+  int my_vit = int(m_organism->GetVitality() + 0.5);
   m_organism->SetCellData(my_vit);
 
   return true;
@@ -8559,7 +8559,7 @@ bool cHardwareCPU::Inst_GetFacedVitalityDiff(cAvidaContext& ctx)
   
   const int out_reg = FindModifiedRegister(REG_BX);
   // return % diff
-  int vit_diff = (neighbor->GetVitality() -  m_organism->GetVitality())/m_organism->GetVitality() * 100 + 0.5;
+  int vit_diff = int((neighbor->GetVitality() -  m_organism->GetVitality())/m_organism->GetVitality() * 100 + 0.5);
   GetRegister(out_reg) = vit_diff;
   return true;
 }
@@ -8661,8 +8661,8 @@ bool cHardwareCPU::Inst_GetAttackOdds(cAvidaContext& ctx)
   
   int odds_I_dont_die;
   // return odds as %
-  if (attacker_odds > target_odds) odds_I_dont_die = (1 - target_odds) * 100 + 0.5;
-  else odds_I_dont_die = (1 - attacker_odds) * 100 + 0.5;
+  if (attacker_odds > target_odds) odds_I_dont_die = int((1 - target_odds) * 100 + 0.5);
+  else odds_I_dont_die = int((1 - attacker_odds) * 100 + 0.5);
   
   const int out_reg = FindModifiedRegister(REG_BX);
   GetRegister(out_reg) = odds_I_dont_die;
